+++ packages/multichain-account-service/src/types.ts
@@ -1,4 +1,8 @@
 import type {
+  Bip44Account,
+  MultichainAccountGroup,
+} from '@metamask/account-api';
+import type {
   AccountsControllerAccountAddedEvent,
   AccountsControllerAccountRemovedEvent,
   AccountsControllerGetAccountAction,
@@ -6,6 +10,7 @@ import type {
   AccountsControllerListMultichainAccountsAction,
 } from '@metamask/accounts-controller';
 import type { RestrictedMessenger } from '@metamask/base-controller';
+import type { KeyringAccount } from '@metamask/keyring-api';
 import type {
   KeyringControllerGetStateAction,
   KeyringControllerStateChangeEvent,
@@ -88,11 +93,23 @@ export type MultichainAccountServiceActions =
   | MultichainAccountServiceAlignWalletsAction
   | MultichainAccountServiceGetIsAlignmentInProgressAction;
 
+export type MultichainAccountServiceMultichainAccountGroupCreatedEvent = {
+  type: `${typeof serviceName}:multichainAccountGroupCreated`;
+  payload: [MultichainAccountGroup<Bip44Account<KeyringAccount>>];
+};
+
+export type MultichainAccountServiceMultichainAccountGroupUpdatedEvent = {
+  type: `${typeof serviceName}:multichainAccountGroupUpdated`;
+  payload: [MultichainAccountGroup<Bip44Account<KeyringAccount>>];
+};
+
 /**
  * All events that {@link MultichainAccountService} publishes so that other modules
  * can subscribe to them.
  */
-export type MultichainAccountServiceEvents = never;
+export type MultichainAccountServiceEvents =
+  | MultichainAccountServiceMultichainAccountGroupCreatedEvent
+  | MultichainAccountServiceMultichainAccountGroupUpdatedEvent;
 
 /**
  * All actions registered by other modules that {@link MultichainAccountService}